refactor: 스타일 폴더 안 컴포넌트별 폴더 재구분
