+++ src/components/ItemBtn.tsx
@@ -3,7 +3,7 @@ type ItemBtnProps = {
   onClick: () => void;
 };
 
-import { ExpandBtn } from '@/styles/ItemBtn.styles';
+import { ExpandBtn } from '@/styles/Item/ItemBtn.styles';
 
 function ItemBtn({ isExpanded, onClick }: ItemBtnProps) {
   return <ExpandBtn onClick={onClick}>{isExpanded ? '접기' : '더보기'}</ExpandBtn>;
