+++ src/styles/Sex/SexContainer.styles.ts
